Move updated color methods to exisiting Color file
